Extract TestOnSelectedCandidatePairChange
Extract TestOnSelectedCandidatePairChange to its own file.
